refactor(arch): move launch pipeline + status assembly + MethodContext to daemon
Decompose the ipc/methods.rs god-file (3283 -> 1128 lines) and remove the
proxy->ipc inversion at its root: MethodContext is daemon-state, not an IPC
concept, so it relocates with PersistedState/LaunchEnv to daemon/context.rs.
With the hub in the daemon layer, the launch pipeline and status assembly
follow so every edge points downward.

- daemon/context.rs: MethodContext + PersistedState + LaunchEnv.
- daemon/launch_service.rs: compose_and_spawn (was start_model_inner) +
  StartParams/StartedLaunch/LaunchModeWire + launch-exclusive helpers.
- ipc/status.rs: status_response + backends_status document assembly.
- ipc/methods.rs is now thin dispatch + parse/serialize adapters.
- proxy/launch.rs calls daemon::launch_service::compose_and_spawn;
  lemonade_snapshot_id becomes RunningSnapshot::lemonade_backend_id so the
  last proxy->ipc::methods code edge is gone.

Wire shapes byte-stable (status top-level-key golden pins it). 1977 tests
green; no proxy->ipc::methods code edges remain.
