+++ benches/proxy_overhead.rs
@@ -56,6 +56,7 @@ use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
 
 use criterion::{criterion_group, criterion_main, Criterion, Throughput};
 use llamastash::backend::llama_cpp::LlamaCppBackend;
+use llamastash::daemon::context::MethodContext;
 use llamastash::daemon::probe::ProbeOptions;
 use llamastash::daemon::registry::SupervisorRegistry;
 use llamastash::daemon::shutdown::ShutdownToken;
@@ -63,7 +64,6 @@ use llamastash::daemon::supervisor::{spawn as supervisor_spawn, ManagedSpawn, Ma
 use llamastash::discovery::{DiscoveredModel, ModelCatalog, ModelSource};
 use llamastash::gguf::identity::ModelId;
 use llamastash::gguf::metadata::{ModeHint, ModelMetadata, Quant};
-use llamastash::ipc::methods::MethodContext;
 use llamastash::launch::mode::LaunchMode;
 use llamastash::launch::params::LaunchParams;
 use llamastash::proxy::server::{loopback_addr, new_status_cell, serve, ProxyStatus, StatusCell};
