+++ src/ipc/status.rs
@@ -1,0 +1,735 @@
+//! Assembly of the daemon's `status` IPC document.
+//!
+//! `status_response` is the read-only snapshot the `status` method
+//! returns: managed + delegated + external model rows, the host/GPU
+//! samplers, the backend matrix, and the proxy block. The wire shape is
+//! an agent-facing contract — every top-level key and field name here is
+//! pinned by `status_top_level_key_set_is_stable` and the surrounding
+//! `status_*` tests, so changes must stay byte-stable.
+
+use std::sync::atomic::Ordering;
+
+use serde_json::{json, Value};
+
+use crate::config::loader::LemonadeConfig;
+use crate::daemon::context::MethodContext;
+use crate::daemon::host_metrics::HostMetricsSnapshot;
+use crate::ipc::methods::flatten_state;
+
+/// Snapshot every active managed model plus the daemon's GPU info.
+/// `status` is read-only; never triggers any state-machine transitions.
+pub(crate) async fn status_response(ctx: &MethodContext) -> Value {
+  let snap = ctx.supervisors.snapshot().await;
+  // Post-launch actuals live on the persisted running snapshot
+  // (stamped by the recorder on Ready); the live status row is built
+  // from the supervisor, so cross-reference by (id, port) to surface
+  // the resolved context.
+  let running = ctx.state.snapshot().await.running;
+  let mut models: Vec<Value> = Vec::with_capacity(snap.len());
+  for (launch_id, model) in snap {
+    let state = model.state().await;
+    let pid = model.pid().await;
+    let ready_at = model.ready_at().await;
+    // Wrap `ManagedState` in a small `{state, cause?}` object: a
+    // lowercase string label plus an optional sibling `cause`, so an
+    // `Error{cause}` surfaces its cause as a plain field rather than
+    // hidden inside a serde tagged-enum content blob.
+    let state_obj = match state.cause() {
+      Some(cause) => json!({"state": state.label(), "cause": cause}),
+      None => json!({"state": state.label()}),
+    };
+    // `params` so an agent can reproduce the launch without a
+    // separate `last_params_list` call.
+    let params = model.params();
+    let params_json = json!({
+      "model_path": params.model_path,
+      "mode": model.mode().label(),
+      "ctx": params.ctx,
+      "port": params.port,
+      "reasoning": params.reasoning,
+      "knobs": &params.knobs,
+      "extras": params
+        .extras
+        .iter()
+        .map(|s| s.to_string_lossy().into_owned())
+        .collect::<Vec<_>>(),
+    });
+    let latest = model.latest_resource().await;
+    let latest_rss_bytes = latest.as_ref().map(|r| r.rss_bytes);
+    let latest_cpu_pct = latest.as_ref().map(|r| r.cpu_percent);
+    let actuals = running
+      .iter()
+      .find(|r| r.port == model.port())
+      .map(|r| r.actuals);
+    let resolved_ctx = actuals.and_then(|a| a.resolved_ctx);
+    let ctx_clamped = actuals.map(|a| a.ctx_clamped).unwrap_or(false);
+    let row = json!({
+      "launch_id": launch_id,
+      "id": model.id(),
+      "port": model.port(),
+      "mode": model.mode().label(),
+      "pid": pid,
+      "ready_at": ready_at,
+      "state": state_obj,
+      "params": params_json,
+      "latest_rss_bytes": latest_rss_bytes,
+      "latest_cpu_pct": latest_cpu_pct,
+      // Resolved context window `--fit` chose; null until the
+      // post-Ready `/props` fetch lands or when the build omits it.
+      "resolved_ctx": resolved_ctx,
+      // True when `--fit` had to clamp ctx to the floor under memory
+      // pressure (a soft notice); strict mode refuses such launches.
+      "ctx_clamped": ctx_clamped,
+    });
+    models.push(row);
+  }
+  // Delegated Lemonade models — the registry holds only the shared
+  // umbrella (one row whose path is the `lemond` binary), but every
+  // model made resident via `start_model` persisted a RunningSnapshot
+  // at the umbrella's port. Project each as a first-class row: the
+  // synthetic `lemonade://<name>` path matches the catalog entry, so
+  // the TUI list pane and `llamastash list` show the *model* as
+  // running, not just the umbrella. State comes from the preload
+  // task's recorded outcome (`Loading` / `Ready` / `Error{cause}`);
+  // a snapshot with no recorded outcome (re-adopted across a daemon
+  // restart) falls back to mirroring the umbrella. Rows are emitted
+  // only while the umbrella is registered; with it gone the
+  // snapshots are unreachable leftovers (the boot sweep reaps them).
+  if let Some(umbrella) = ctx
+    .supervisors
+    .get(&crate::backend::lemonade::umbrella_launch_id())
+    .await
+  {
+    let ustate_obj = flatten_state(&umbrella.state().await);
+    let upid = umbrella.pid().await;
+    for running_snap in ctx.state.snapshot().await.running.iter() {
+      let Some(backend_id) = running_snap.lemonade_backend_id() else {
+        continue;
+      };
+      let state_obj = match ctx.supervisors.delegated_state(&backend_id.name).await {
+        Some(s) => flatten_state(&s),
+        None => ustate_obj.clone(),
+      };
+      let synthetic_id = crate::gguf::identity::ModelId {
+        path: running_snap.params.model_path.clone(),
+        header_blake3: [0u8; 32],
+      };
+      let params_json = json!({
+        "model_path": running_snap.params.model_path,
+        "mode": running_snap.params.mode.label(),
+        "ctx": running_snap.params.ctx,
+        "port": running_snap.params.port,
+        "reasoning": running_snap.params.reasoning,
+        "knobs": &running_snap.params.knobs,
+        "extras": running_snap.params.extras
+          .iter()
+          .map(|s| s.to_string_lossy().into_owned())
+          .collect::<Vec<_>>(),
+      });
+      models.push(json!({
+        "launch_id": crate::backend::lemonade::delegated_launch_id(&backend_id.name),
+        "id": synthetic_id,
+        "port": running_snap.port,
+        "mode": running_snap.params.mode.label(),
+        "pid": upid,
+        "ready_at": running_snap.started_at,
+        "state": state_obj,
+        "params": params_json,
+        // Resource readings stay on the umbrella's own row — mirroring
+        // its RSS onto every resident model would double-count it.
+        "latest_rss_bytes": Value::Null,
+        "latest_cpu_pct": Value::Null,
+      }));
+    }
+  }
+  // External — read-only rows for `llama-server` processes the
+  // daemon doesn't own. Populated by the startup orphan sweep.
+  // Stable shape: `{pid, cmdline, model_path, port, launched_by_llamastash}`.
+  let external_snapshot = ctx.external.read().await.clone();
+  let external: Vec<Value> = external_snapshot
+    .iter()
+    .map(|e| {
+      json!({
+        "pid": e.pid,
+        "cmdline": e.cmdline,
+        "model_path": e.model_path,
+        // Tier-A orphan-tracking fields. `port` lets agents diff the
+        // wire snapshot against `ss`/`lsof` without parsing argv;
+        // `launched_by_llamastash` exposes whether the orphan carries
+        // the supervisor's spawn marker so operators can spot
+        // sibling-instance orphans at a glance.
+        "port": e.port,
+        "launched_by_llamastash": e.launched_by_llamastash,
+      })
+    })
+    .collect();
+  // Host-level metrics (CPU%, RAM, GPU util/temp/VRAM aggregates).
+  // Sampled by the daemon's `host_metrics` task at 1 Hz. When no
+  // sampler is attached (catalog-only contexts), emit a default
+  // snapshot rather than `null` so clients see a stable object
+  // shape — `gpu_backend == "unsampled"` already distinguishes the
+  // never-sampled case from a real reading.
+  //
+  // Serialize the snapshot directly under the read lock instead of
+  // cloning it out first; `HostMetricsSnapshot` already implements
+  // `Serialize` for `&Self`, so this saves one full struct clone
+  // (including the `gpu_backend: String`) per status call.
+  let host = match &ctx.host_metrics {
+    Some(slot) => {
+      let host_snap = slot.read().await;
+      serde_json::to_value(&*host_snap).unwrap_or(Value::Null)
+    }
+    None => {
+      let default_snap = HostMetricsSnapshot {
+        gpu_backend: HostMetricsSnapshot::UNINITIALIZED_BACKEND.into(),
+        ..HostMetricsSnapshot::default()
+      };
+      serde_json::to_value(default_snap).unwrap_or(Value::Null)
+    }
+  };
+  // Prefer the live GpuInfo cell when the sampler is attached so
+  // `status.gpu` follows hotplug / late-driver-load events. Falls
+  // back to the boot-time `ctx.gpu` snapshot when the sampler is
+  // off (catalog-only tests).
+  let gpu = match &ctx.gpu_live {
+    Some(slot) => serde_json::to_value(&*slot.read().await).unwrap_or(Value::Null),
+    None => serde_json::to_value(ctx.gpu.as_ref()).unwrap_or(Value::Null),
+  };
+  // Proxy block — read-only projection of the listener's shared
+  // status cell. Catalog-only tests that never bring the proxy up
+  // leave `proxy_status` as `None`; the field is omitted in that
+  // case so those fixtures stay byte-identical. The wire shape is:
+  //
+  // ```
+  // "proxy": {
+  //   "enabled": bool,
+  //   "listen": "127.0.0.1:11434" | null,
+  //   "status": "disabled" | "listening" | "port_in_use" | "unbound",
+  //   "bind_error": "permission denied" | null,
+  // }
+  // ```
+  let proxy = ctx.proxy_status.as_ref().map(project_proxy_status);
+  // Launch device catalog — the exact `--device` selectors the TUI
+  // picker may offer, each tagged with the binary that owns it.
+  // Sourced from every configured binary's `--list-devices` (not from
+  // vendor tools), so what the picker shows is precisely what
+  // `llama-server` will accept. Empty when no binary is configured.
+  let device_catalog = match ctx.launch.as_ref() {
+    Some(env) => {
+      let catalog_snap = env.device_catalog.read().await;
+      serde_json::to_value(&*catalog_snap).unwrap_or(Value::Null)
+    }
+    None => Value::Null,
+  };
+  let backends = backends_status(ctx).await;
+  let mut body = json!({
+    "models": models,
+    "external": external,
+    "gpu": gpu,
+    "host": host,
+    "device_catalog": device_catalog,
+    "backends": backends,
+    "daemon": {
+      "pid": std::process::id(),
+      "uptime_seconds": ctx.started_at.elapsed().as_secs(),
+      "active_connections": ctx.active_connections.load(Ordering::Relaxed),
+      "build": env!("CARGO_PKG_VERSION"),
+      "server_path": ctx
+        .launch
+        .as_ref()
+        .map(|env| env.binary.display().to_string()),
+      "ipc_url": ctx.ipc_url,
+    },
+  });
+  if let Some(proxy) = proxy {
+    if let Some(obj) = body.as_object_mut() {
+      obj.insert("proxy".into(), proxy);
+    }
+  }
+  body
+}
+
+/// Build the `status.backends` array: one row per backend with
+/// whether its binary is installed on this host and which accelerators it
+/// can run on. llama.cpp's accelerator set unions its CPU floor with the
+/// GPU backends the live device catalog reveals; Lemonade reports its
+/// static cpu+npu (a live `lemond` system-info probe is deferred).
+async fn backends_status(ctx: &MethodContext) -> Value {
+  use crate::backend::lemonade::LemonadeBackend;
+  use crate::backend::llama_cpp::LlamaCppBackend;
+  use crate::backend::Backend;
+
+  let llama = LlamaCppBackend::new();
+  let llama_installed = ctx
+    .launch
+    .as_ref()
+    .map(|e| e.binary.exists())
+    .unwrap_or(false);
+  let mut llama_acc = llama.accelerators();
+  if let Some(env) = ctx.launch.as_ref() {
+    let cat = env.device_catalog.read().await;
+    for d in cat.iter() {
+      if let Some(a) = accelerator_from_selector(&d.selector) {
+        llama_acc.insert(a);
+      }
+    }
+  }
+
+  let lemonade = LemonadeBackend::new();
+  let mut lemonade_row = backend_row(
+    lemonade.id(),
+    lemonade.lifecycle().label(),
+    lemond_installed(&ctx.lemonade),
+    lemonade.accelerators().labels(),
+  );
+  // Managed-multiplexer health: surface whether the umbrella llamastash
+  // supervises is actually up, so `status` distinguishes "installed" (binary on
+  // disk) from "running" (umbrella Ready). When enabled but down — e.g. the
+  // configured port was already taken at boot — this reads `not running`, which
+  // is the only visible signal outside the daemon log.
+  if let Some(obj) = lemonade_row.as_object_mut() {
+    obj.insert("enabled".into(), json!(ctx.lemonade.enabled));
+    obj.insert("umbrella".into(), json!(lemonade_umbrella_state(ctx).await));
+  }
+  // Each row carries its resolved `binary` path when one exists, so
+  // clients (the TUI's Daemon panel server row) can list every enabled
+  // backend's executable generically — no per-backend client code.
+  let mut llama_row = backend_row(
+    llama.id(),
+    llama.lifecycle().label(),
+    llama_installed,
+    llama_acc.labels(),
+  );
+  let llama_binary = ctx.launch.as_ref().map(|e| e.binary.display().to_string());
+  set_backend_binary(&mut llama_row, llama_binary);
+  let lemonade_binary =
+    crate::backend::lemonade::resolve_lemond_binary(&ctx.lemonade).map(|b| b.display().to_string());
+  set_backend_binary(&mut lemonade_row, lemonade_binary);
+  json!([llama_row, lemonade_row])
+}
+
+/// Attach a backend row's resolved `binary` path; absent when no binary
+/// resolves so clients can tell "backend known" from "executable found".
+fn set_backend_binary(row: &mut Value, binary: Option<String>) {
+  if let (Some(obj), Some(bin)) = (row.as_object_mut(), binary) {
+    obj.insert("binary".into(), json!(bin));
+  }
+}
+
+/// The managed `lemond` umbrella's state for `status`, distinct from the
+/// `installed` (binary-resolvable) signal. `disabled` when the backend is off;
+/// otherwise reflects the supervised umbrella: `running` (Ready), `starting`
+/// (spawned, probing), or `not running` (never came up / exited — commonly a
+/// boot-time port conflict).
+async fn lemonade_umbrella_state(ctx: &MethodContext) -> &'static str {
+  use crate::daemon::supervisor::ManagedState;
+  if !ctx.lemonade.enabled {
+    return "disabled";
+  }
+  match ctx
+    .supervisors
+    .get(&crate::backend::lemonade::umbrella_launch_id())
+    .await
+  {
+    Some(m) => match m.state().await {
+      ManagedState::Ready => "running",
+      ManagedState::Launching | ManagedState::Loading => "starting",
+      ManagedState::Error { .. } | ManagedState::Stopping | ManagedState::Stopped => "not running",
+    },
+    None => "not running",
+  }
+}
+
+fn backend_row(id: &str, lifecycle: &str, installed: bool, accelerators: Vec<&str>) -> Value {
+  json!({
+    "id": id,
+    "lifecycle": lifecycle,
+    "installed": installed,
+    "accelerators": accelerators,
+  })
+}
+
+/// Map a llama.cpp `--device` selector prefix to an accelerator class.
+fn accelerator_from_selector(selector: &str) -> Option<crate::backend::Accelerator> {
+  use crate::backend::Accelerator;
+  let s = selector.to_ascii_lowercase();
+  if s.starts_with("cuda") {
+    Some(Accelerator::Cuda)
+  } else if s.starts_with("rocm") {
+    Some(Accelerator::Rocm)
+  } else if s.starts_with("vulkan") {
+    Some(Accelerator::Vulkan)
+  } else if s.starts_with("metal") {
+    Some(Accelerator::Metal)
+  } else {
+    None
+  }
+}
+
+/// The "installed" signal for the Lemonade backend in `status`. Honors the
+/// full resolution order — explicit `lemonade.binary` first, then `lemond` /
+/// `lemonade` on PATH — so a user who points at an off-PATH `lemond` still
+/// reads as installed.
+fn lemond_installed(cfg: &LemonadeConfig) -> bool {
+  crate::backend::lemonade::resolve_lemond_binary(cfg).is_some()
+}
+
+/// Project the proxy listener's status cell into the wire shape
+/// surfaced under `status.proxy`. The cell is the single source of
+/// truth — the listener task writes every transition
+/// (Disabled / Listening / PortInUse / Unbound) and this is the
+/// read side.
+///
+/// `listen` is the *attempted* address: `Disabled` emits `null`
+/// (no bind was attempted), every other variant carries the address
+/// the daemon tried to bind. `bind_error` is non-null only when the
+/// variant is `Unbound` — `PortInUse` is its own discriminator and
+/// callers shouldn't need a parallel string to recognise it.
+fn project_proxy_status(cell: &crate::proxy::StatusCell) -> Value {
+  use crate::proxy::ProxyStatus;
+  let snapshot = cell.read().unwrap_or_else(|e| e.into_inner()).clone();
+  match snapshot {
+    ProxyStatus::Disabled => json!({
+      "enabled": false,
+      "listen": Value::Null,
+      "host": Value::Null,
+      "status": "disabled",
+      "auth": "none",
+      "bind_error": Value::Null,
+    }),
+    ProxyStatus::Listening {
+      addr,
+      auth_enforced,
+    } => json!({
+      "enabled": true,
+      "listen": addr.to_string(),
+      "host": addr.ip().to_string(),
+      "status": "listening",
+      "auth": if auth_enforced { "enforced" } else { "none" },
+      "bind_error": Value::Null,
+    }),
+    ProxyStatus::PortInUse { addr } => json!({
+      "enabled": true,
+      "listen": addr.to_string(),
+      "host": addr.ip().to_string(),
+      "status": "port_in_use",
+      "auth": "none",
+      "bind_error": Value::Null,
+    }),
+    ProxyStatus::Unbound { addr, bind_error } => json!({
+      "enabled": true,
+      "listen": addr.to_string(),
+      "host": addr.ip().to_string(),
+      "status": "unbound",
+      "auth": "none",
+      "bind_error": bind_error,
+    }),
+    // Refused to expose a non-loopback proxy without auth. The daemon
+    // is healthy; the proxy just didn't bind. `auth` reports
+    // `"required"` so a client can distinguish this from a plain bind
+    // failure and surface the fix.
+    ProxyStatus::RefusedInsecure { addr } => json!({
+      "enabled": true,
+      "listen": addr.to_string(),
+      "host": addr.ip().to_string(),
+      "status": "refused_insecure",
+      "auth": "required",
+      "bind_error":
+        "refused to bind a non-loopback proxy without authentication; \
+         set proxy.api_key or pass --insecure-no-auth",
+    }),
+  }
+}
+
+#[cfg(test)]
+mod tests {
+  use std::path::PathBuf;
+
+  use serde_json::{json, Value};
+
+  use crate::daemon::context::MethodContext;
+  use crate::daemon::shutdown::ShutdownToken;
+  use crate::ipc::methods::dispatch_request;
+  use crate::ipc::protocol::Request;
+  use crate::launch::mode::LaunchMode;
+  use crate::launch::params::LaunchParams;
+
+  fn ctx() -> MethodContext {
+    MethodContext::new(ShutdownToken::new())
+  }
+
+  /// A delegated-lemonade snapshot the way `start_delegated_lemonade`
+  /// persists one: Backend identity + the synthetic `lemonade://` path.
+  fn lemonade_running_snapshot(
+    name: &str,
+    port: u16,
+  ) -> crate::daemon::state_store::RunningSnapshot {
+    crate::daemon::state_store::RunningSnapshot {
+      id: crate::backend::identity::ModelIdentity::Backend(
+        crate::backend::identity::BackendModelId {
+          backend: crate::backend::lemonade::LEMONADE_BACKEND_ID.to_string(),
+          name: name.to_string(),
+        },
+      ),
+      pid: 0,
+      port,
+      started_at: 0,
+      params: LaunchParams::new(
+        PathBuf::from(format!("lemonade://{name}")),
+        LaunchMode::Chat,
+      ),
+      actuals: Default::default(),
+    }
+  }
+
+  #[tokio::test]
+  async fn status_includes_daemon_health_block() {
+    let c = ctx();
+    let resp = dispatch_request(&c, Request::new(1, "status", None)).await;
+    let body = resp.result.expect("status result");
+    let daemon = body
+      .get("daemon")
+      .expect("status must include daemon health block");
+    assert!(daemon["pid"].is_number());
+    assert!(daemon["uptime_seconds"].is_number());
+    assert_eq!(daemon["active_connections"], json!(0));
+  }
+
+  /// Agent-facing contract guard: pins the exact top-level key set of a
+  /// catalog-only `status` response so the status-assembly extraction
+  /// stays byte-stable (`proxy` is omitted when the cell is absent).
+  #[tokio::test]
+  async fn status_top_level_key_set_is_stable() {
+    let c = ctx();
+    let resp = dispatch_request(&c, Request::new(1, "status", None)).await;
+    let body = resp.result.expect("status result");
+    let mut keys: Vec<&str> = body
+      .as_object()
+      .expect("status is a JSON object")
+      .keys()
+      .map(String::as_str)
+      .collect();
+    keys.sort_unstable();
+    assert_eq!(
+      keys,
+      vec![
+        "backends",
+        "daemon",
+        "device_catalog",
+        "external",
+        "gpu",
+        "host",
+        "models",
+      ],
+      "catalog-only status top-level keys drifted"
+    );
+  }
+
+  #[tokio::test]
+  async fn status_includes_backends_block() {
+    let c = ctx();
+    let resp = dispatch_request(&c, Request::new(1, "status", None)).await;
+    let body = resp.result.expect("status result");
+    let backends = body
+      .get("backends")
+      .and_then(|v| v.as_array())
+      .expect("status must include a backends array");
+    let ids: Vec<&str> = backends
+      .iter()
+      .filter_map(|b| b.get("id").and_then(|v| v.as_str()))
+      .collect();
+    assert!(
+      ids.contains(&"llamacpp"),
+      "backends must list llamacpp: {ids:?}"
+    );
+    assert!(
+      ids.contains(&"lemonade"),
+      "backends must list lemonade: {ids:?}"
+    );
+    // Each row carries the installed/lifecycle/accelerator fields;
+    // llama.cpp always offers CPU.
+    let llama = backends
+      .iter()
+      .find(|b| b["id"] == "llamacpp")
+      .expect("llamacpp row");
+    assert!(llama["installed"].is_boolean());
+    assert_eq!(llama["lifecycle"], json!("process_per_model"));
+    let accel: Vec<&str> = llama["accelerators"]
+      .as_array()
+      .unwrap()
+      .iter()
+      .filter_map(|v| v.as_str())
+      .collect();
+    assert!(accel.contains(&"cpu"), "llama.cpp floor is cpu: {accel:?}");
+    // The Lemonade row is a managed-multiplexer offering cpu+npu.
+    let lemon = backends
+      .iter()
+      .find(|b| b["id"] == "lemonade")
+      .expect("lemonade row");
+    assert!(lemon["installed"].is_boolean());
+    assert_eq!(lemon["lifecycle"], json!("managed_multiplexer"));
+    let lacc: Vec<&str> = lemon["accelerators"]
+      .as_array()
+      .unwrap()
+      .iter()
+      .filter_map(|v| v.as_str())
+      .collect();
+    assert!(lacc.contains(&"npu"), "lemonade offers npu: {lacc:?}");
+  }
+
+  #[tokio::test]
+  async fn status_omits_proxy_block_when_cell_is_absent() {
+    // Catalog-only contexts (`MethodContext::new`) leave
+    // `proxy_status` as `None`. The wire shape must omit the
+    // `proxy` field entirely so such fixtures stay byte-identical —
+    // callers that don't surface a proxy don't get a confusing
+    // `"proxy": null` blob either.
+    let c = ctx();
+    let resp = dispatch_request(&c, Request::new(1, "status", None)).await;
+    let body = resp.result.expect("status result");
+    assert!(
+      body.get("proxy").is_none(),
+      "proxy block must be absent when no cell is attached: {body}"
+    );
+  }
+
+  #[tokio::test]
+  async fn status_emits_proxy_listening_block() {
+    use crate::proxy;
+    use std::net::SocketAddr;
+    let addr: SocketAddr = "127.0.0.1:11434".parse().unwrap();
+    let cell = proxy::server::new_status_cell();
+    *cell.write().unwrap() = proxy::ProxyStatus::Listening {
+      addr,
+      auth_enforced: false,
+    };
+    let c = MethodContext::new(ShutdownToken::new()).with_proxy_status(cell);
+    let resp = dispatch_request(&c, Request::new(1, "status", None)).await;
+    let body = resp.result.expect("status result");
+    let proxy = body.get("proxy").expect("proxy block present");
+    assert_eq!(proxy["enabled"], json!(true));
+    assert_eq!(proxy["listen"], json!("127.0.0.1:11434"));
+    assert_eq!(proxy["host"], json!("127.0.0.1"));
+    assert_eq!(proxy["status"], json!("listening"));
+    assert_eq!(proxy["auth"], json!("none"));
+    assert_eq!(proxy["bind_error"], Value::Null);
+  }
+
+  #[tokio::test]
+  async fn status_listening_reports_auth_enforced_and_lan_host() {
+    use crate::proxy;
+    use std::net::SocketAddr;
+    let addr: SocketAddr = "0.0.0.0:11434".parse().unwrap();
+    let cell = proxy::server::new_status_cell();
+    *cell.write().unwrap() = proxy::ProxyStatus::Listening {
+      addr,
+      auth_enforced: true,
+    };
+    let c = MethodContext::new(ShutdownToken::new()).with_proxy_status(cell);
+    let resp = dispatch_request(&c, Request::new(1, "status", None)).await;
+    let proxy = resp.result.expect("status result");
+    let proxy = proxy.get("proxy").expect("proxy block present");
+    assert_eq!(proxy["host"], json!("0.0.0.0"));
+    assert_eq!(proxy["auth"], json!("enforced"));
+  }
+
+  #[tokio::test]
+  async fn status_omits_delegated_lemonade_rows_without_umbrella() {
+    // A snapshot with no registered umbrella is an unreachable leftover
+    // (umbrella crashed / was stopped): emitting a row for it would
+    // offer a stop affordance against nothing. The happy path (umbrella
+    // up → rows emitted) is covered in `lemonade_umbrella_test.rs`.
+    let c = ctx();
+    c.state
+      .mutate(|s| s.running.push(lemonade_running_snapshot("Qwen-X", 13305)))
+      .await;
+    let resp = dispatch_request(&c, Request::new(1, "status", None)).await;
+    let body = resp.result.expect("status result");
+    let models = body["models"].as_array().expect("models array");
+    assert!(
+      !models.iter().any(|m| m["launch_id"]
+        .as_str()
+        .unwrap_or("")
+        .starts_with("lemonade:")),
+      "no delegated rows without a registered umbrella: {models:?}"
+    );
+  }
+
+  #[tokio::test]
+  async fn status_emits_proxy_refused_insecure_block() {
+    use crate::proxy;
+    use std::net::SocketAddr;
+    let addr: SocketAddr = "0.0.0.0:11434".parse().unwrap();
+    let cell = proxy::server::new_status_cell();
+    *cell.write().unwrap() = proxy::ProxyStatus::RefusedInsecure { addr };
+    let c = MethodContext::new(ShutdownToken::new()).with_proxy_status(cell);
+    let resp = dispatch_request(&c, Request::new(1, "status", None)).await;
+    let proxy = resp.result.expect("status result");
+    let proxy = proxy.get("proxy").expect("proxy block present");
+    assert_eq!(proxy["status"], json!("refused_insecure"));
+    assert_eq!(proxy["auth"], json!("required"));
+    assert_eq!(proxy["host"], json!("0.0.0.0"));
+    assert!(
+      proxy["bind_error"]
+        .as_str()
+        .unwrap()
+        .contains("--insecure-no-auth"),
+      "refused_insecure must explain the fix: {proxy}"
+    );
+  }
+
+  #[tokio::test]
+  async fn status_emits_proxy_disabled_block() {
+    use crate::proxy;
+    let cell = proxy::server::new_status_cell();
+    // `new_status_cell` already seeds `Disabled`.
+    let c = MethodContext::new(ShutdownToken::new()).with_proxy_status(cell);
+    let resp = dispatch_request(&c, Request::new(1, "status", None)).await;
+    let body = resp.result.expect("status result");
+    let proxy = body.get("proxy").expect("proxy block present");
+    assert_eq!(proxy["enabled"], json!(false));
+    assert_eq!(proxy["listen"], Value::Null);
+    assert_eq!(proxy["status"], json!("disabled"));
+    assert_eq!(proxy["bind_error"], Value::Null);
+  }
+
+  #[tokio::test]
+  async fn status_emits_proxy_port_in_use_block() {
+    use crate::proxy;
+    use std::net::SocketAddr;
+    let addr: SocketAddr = "127.0.0.1:11434".parse().unwrap();
+    let cell = proxy::server::new_status_cell();
+    *cell.write().unwrap() = proxy::ProxyStatus::PortInUse { addr };
+    let c = MethodContext::new(ShutdownToken::new()).with_proxy_status(cell);
+    let resp = dispatch_request(&c, Request::new(1, "status", None)).await;
+    let body = resp.result.expect("status result");
+    let proxy = body.get("proxy").expect("proxy block present");
+    assert_eq!(proxy["enabled"], json!(true));
+    assert_eq!(proxy["listen"], json!("127.0.0.1:11434"));
+    assert_eq!(proxy["status"], json!("port_in_use"));
+    // PortInUse is its own discriminator; no parallel bind_error
+    // string — the wire shape pins this so parsers don't have to
+    // double-check.
+    assert_eq!(proxy["bind_error"], Value::Null);
+  }
+
+  #[tokio::test]
+  async fn status_emits_proxy_unbound_block_with_bind_error() {
+    use crate::proxy;
+    use std::net::SocketAddr;
+    let addr: SocketAddr = "127.0.0.1:80".parse().unwrap();
+    let cell = proxy::server::new_status_cell();
+    *cell.write().unwrap() = proxy::ProxyStatus::Unbound {
+      addr,
+      bind_error: "permission denied".to_string(),
+    };
+    let c = MethodContext::new(ShutdownToken::new()).with_proxy_status(cell);
+    let resp = dispatch_request(&c, Request::new(1, "status", None)).await;
+    let body = resp.result.expect("status result");
+    let proxy = body.get("proxy").expect("proxy block present");
+    assert_eq!(proxy["enabled"], json!(true));
+    assert_eq!(proxy["listen"], json!("127.0.0.1:80"));
+    assert_eq!(proxy["status"], json!("unbound"));
+    assert_eq!(proxy["bind_error"], json!("permission denied"));
+  }
+}
