+++ tests/proxy_ui.rs
@@ -18,6 +18,7 @@ use std::{
 };
 
 use llamastash::backend::llama_cpp::LlamaCppBackend;
+use llamastash::daemon::context::MethodContext;
 use llamastash::daemon::probe::ProbeOptions;
 use llamastash::daemon::registry::{LaunchId, SupervisorRegistry};
 use llamastash::daemon::shutdown::ShutdownToken;
@@ -25,7 +26,6 @@ use llamastash::daemon::supervisor::{spawn as supervisor_spawn, ManagedSpawn, Ma
 use llamastash::discovery::{DiscoveredModel, ModelCatalog, ModelSource};
 use llamastash::gguf::identity::ModelId;
 use llamastash::gguf::metadata::{ModeHint, ModelMetadata, Quant};
-use llamastash::ipc::methods::MethodContext;
 use llamastash::launch::mode::LaunchMode;
 use llamastash::launch::params::LaunchParams;
 use llamastash::proxy::server::{loopback_addr, new_status_cell, serve, ProxyStatus, StatusCell};
